+++ x/inflation/types/genesis.go
@@ -1,5 +1,7 @@
 package types
 
+import "fmt"
+
 // DefaultGenesis returns the default inflation genesis state
 func DefaultGenesis() *GenesisState {
 	return &GenesisState{
@@ -12,6 +14,10 @@ func DefaultGenesis() *GenesisState {
 func (gs GenesisState) Validate() error {
 	if err := gs.Params.Validate(); err != nil {
 		return err
+	}
+
+	if !gs.Params.DecayRate.IsPositive() && gs.DecayEpochStart != 0 {
+		return fmt.Errorf("decay_epoch_start must be zero when decay is disabled")
 	}
 
 	return nil